Add Cinnamon Stars
I changed it from saying to use a blender to beat the egg whites to using an electric mixer. We tried using a blender in 2021 but it didn't reall work. I'm not sure why this one said to use a blender but the chocolate platzen said to use a mixer.